+++ src/main/java/com/dnsimple/data/Certificate.java
@@ -6,8 +6,6 @@ import java.util.List;
 public class Certificate {
     private final Long id;
     private final Long domainId;
-    @Deprecated
-    private final Long contactId;
     private final String commonName;
     private final List<String> alternateNames;
     private final Integer years;
@@ -22,27 +20,6 @@ public class Certificate {
     public Certificate(Long id, Long domainId, String commonName, List<String> alternateNames, Integer years, String state, String authorityIdentifier, Boolean autoRenew, OffsetDateTime createdAt, OffsetDateTime updatedAt, OffsetDateTime expiresAt, String csr) {
         this.id = id;
         this.domainId = domainId;
-        this.contactId = null;
-        this.commonName = commonName;
-        this.alternateNames = alternateNames;
-        this.years = years;
-        this.state = state;
-        this.authorityIdentifier = authorityIdentifier;
-        this.autoRenew = autoRenew;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-        this.expiresAt = expiresAt;
-        this.csr = csr;
-    }
-
-    /**
-     * @deprecated Please, use a constructor without deprecated arguments
-     */
-    @Deprecated
-    public Certificate(Long id, Long domainId, @Deprecated Long contactId, String commonName, List<String> alternateNames, Integer years, String state, String authorityIdentifier, Boolean autoRenew, OffsetDateTime createdAt, OffsetDateTime updatedAt, OffsetDateTime expiresAt, String csr) {
-        this.id = id;
-        this.domainId = domainId;
-        this.contactId = contactId;
         this.commonName = commonName;
         this.alternateNames = alternateNames;
         this.years = years;
@@ -61,11 +38,6 @@ public class Certificate {
 
     public Long getDomainId() {
         return domainId;
-    }
-
-    @Deprecated
-    public Long getContactId() {
-        return contactId;
     }
 
     public String getCommonName() {